# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "K7 E-Library API",
        "description": "Digital library management system API for educational institutions",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8747/api/v1"
        }
    ],
    "paths": {
        "/admin/bulk-upload": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Bulk upload content via ZIP",
                "operationId": "350a3872ad40c3a93652c17f87f65a5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "ZIP file containing PDFs",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bulk upload results"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/content": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List all content (admin)",
                "operationId": "8a5f8bb4c3c728216deb2c8b350475a7",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "under_review",
                                "published",
                                "rejected"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Content list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create content",
                "operationId": "9e45c31a4427a1fe71a16e2c60b3aa00",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "title",
                                    "access_level"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "ebook",
                                            "journal",
                                            "student_project",
                                            "lecture"
                                        ]
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Introduction to Laravel"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "abstract": {
                                        "type": "string"
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "published_year": {
                                        "type": "integer",
                                        "example": 2024
                                    },
                                    "access_level": {
                                        "type": "string",
                                        "enum": [
                                            "public",
                                            "authenticated",
                                            "faculty_only",
                                            "admin_only"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Content created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/content/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update content",
                "operationId": "dd8c226141c60c918561eebf507e87b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "under_review",
                                            "published",
                                            "rejected"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Content updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete content",
                "operationId": "7ac5acd39960e2cc6b940e49112f44d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Content deleted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/content/{id}/publish": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Publish content",
                "operationId": "bca618a02346d9287eee47ec4eaadcf0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Content published"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}